refactor: moved common scm segment code into base type
git and plastic share some common methods and status properties. So moving them in a base type keeps the code base DRY
