+++ src/segment_plastic.go
@@ -6,27 +6,37 @@ import (
 	"strings"
 )
 
+type PlasticStatus struct {
+	ScmStatus
+}
+
+func (s *PlasticStatus) add(code string) {
+	switch code {
+	case "LD":
+		s.Deleted++
+	case "AD", "PR":
+		s.Added++
+	case "LM":
+		s.Moved++
+	case "CH":
+		s.Modified++
+	}
+}
+
 type plastic struct {
-	props properties
-	env   environmentInfo
+	scm
 
-	Moved    int
-	Deleted  int
-	Added    int
-	Modified int
-	Unmerged int
-	Changed  bool
-
+	Status   *PlasticStatus
 	Behind   bool
 	Selector string
 
 	plasticWorkspaceFolder string // root folder of workspace
 }
 
-const (
-	// FullBranchPath displays the full path of a branch
-	FullBranchPath Property = "full_branch_path"
-)
+func (p *plastic) init(props properties, env environmentInfo) {
+	p.props = props
+	p.env = env
+}
 
 func (p *plastic) enabled() bool {
 	if !p.env.hasCommand("cm") {
@@ -48,21 +58,12 @@ func (p *plastic) enabled() bool {
 	return false
 }
 
-func (p *plastic) shouldIgnoreRootRepository(rootDir string) bool {
-	value, ok := p.props[ExcludeFolders]
-	if !ok {
-		return false
-	}
-	excludedFolders := parseStringArray(value)
-	return dirMatchesOneOf(p.env, rootDir, excludedFolders)
-}
-
 func (p *plastic) string() string {
 	displayStatus := p.props.getOneOfBool(FetchStatus, DisplayStatus, false)
 
-	p.Selector = p.getSelector()
+	p.setSelector()
 	if displayStatus {
-		p.getPlasticStatus()
+		p.setPlasticStatus()
 	}
 
 	// use template if available
@@ -88,23 +89,7 @@ func (p *plastic) templateString(segmentTemplate string) string {
 	return text
 }
 
-func (p *plastic) Status() string {
-	var status string
-	stringIfValue := func(value int, prefix string) string {
-		if value > 0 {
-			return fmt.Sprintf(" %s%d", prefix, value)
-		}
-		return ""
-	}
-	status += stringIfValue(p.Added, "+")
-	status += stringIfValue(p.Modified, "~")
-	status += stringIfValue(p.Deleted, "-")
-	status += stringIfValue(p.Moved, ">")
-	status += stringIfValue(p.Unmerged, "x")
-	return strings.TrimSpace(status)
-}
-
-func (p *plastic) getPlasticStatus() {
+func (p *plastic) setPlasticStatus() {
 	output := p.getCmCommandOutput("status", "--all", "--machinereadable")
 	splittedOutput := strings.Split(output, "\n")
 	// compare to head
@@ -113,6 +98,7 @@ func (p *plastic) getPlasticStatus() {
 	p.Behind = headChangeset > currentChangeset
 
 	// parse file state
+	p.Status = &PlasticStatus{}
 	p.parseFilesStatus(splittedOutput)
 }
 
@@ -124,23 +110,9 @@ func (p *plastic) parseFilesStatus(output []string) {
 		if len(line) < 3 {
 			continue
 		}
-		if matchString(`(?i)merge\s+from\s+[0-9]+\s*$`, line) {
-			p.Unmerged++
-			continue
-		}
-		code := line[0:2]
-		switch code {
-		case "LD":
-			p.Deleted++
-		case "AD", "PR":
-			p.Added++
-		case "LM":
-			p.Moved++
-		case "CH":
-			p.Modified++
-		}
+		code := line[:2]
+		p.Status.add(code)
 	}
-	p.Changed = p.Added > 0 || p.Deleted > 0 || p.Modified > 0 || p.Moved > 0 || p.Unmerged > 0
 }
 
 func (p *plastic) parseStringPattern(output, pattern, name string) string {
@@ -169,29 +141,27 @@ func (p *plastic) getHeadChangeset() int {
 	return p.parseIntPattern(output, `\bcs:(?P<cs>[0-9]+?)\s`, "cs", 0)
 }
 
-func (p *plastic) getPlasticFileContents(file string) string {
-	return strings.Trim(p.env.getFileContent(p.plasticWorkspaceFolder+"/.plastic/"+file), " \r\n")
-}
-
-func (p *plastic) getSelector() string {
+func (p *plastic) setSelector() {
 	var ref string
-	selector := p.getPlasticFileContents("plastic.selector")
+	selector := p.getFileContents(p.plasticWorkspaceFolder+"/.plastic/", "plastic.selector")
 	// changeset
 	ref = p.parseChangesetSelector(selector)
 	if ref != "" {
-		return fmt.Sprintf("%s%s", p.props.getString(CommitIcon, "\uF417"), ref)
+		p.Selector = fmt.Sprintf("%s%s", p.props.getString(CommitIcon, "\uF417"), ref)
+		return
 	}
 	// fallback to label
 	ref = p.parseLabelSelector(selector)
 	if ref != "" {
-		return fmt.Sprintf("%s%s", p.props.getString(TagIcon, "\uF412"), ref)
+		p.Selector = fmt.Sprintf("%s%s", p.props.getString(TagIcon, "\uF412"), ref)
+		return
 	}
 	// fallback to branch/smartbranch
 	ref = p.parseBranchSelector(selector)
 	if ref != "" {
 		ref = p.truncateBranch(ref)
 	}
-	return fmt.Sprintf("%s%s", p.props.getString(BranchIcon, "\uE0A0"), ref)
+	p.Selector = fmt.Sprintf("%s%s", p.props.getString(BranchIcon, "\uE0A0"), ref)
 }
 
 func (p *plastic) parseChangesetSelector(selector string) string {
@@ -206,26 +176,7 @@ func (p *plastic) parseBranchSelector(selector string) string {
 	return p.parseStringPattern(selector, `branch "(?P<branch>[\/a-zA-Z0-9\-\_]+?)"`, "branch")
 }
 
-func (p *plastic) init(props properties, env environmentInfo) {
-	p.props = props
-	p.env = env
-}
-
-func (p *plastic) truncateBranch(branch string) string {
-	fullBranchPath := p.props.getBool(FullBranchPath, false)
-	maxLength := p.props.getInt(BranchMaxLength, 0)
-	if !fullBranchPath && len(branch) > 0 {
-		index := strings.LastIndex(branch, "/")
-		branch = branch[index+1:]
-	}
-	if maxLength == 0 || len(branch) <= maxLength {
-		return branch
-	}
-	symbol := p.props.getString(TruncateSymbol, "")
-	return branch[0:maxLength] + symbol
-}
-
-func (p *plastic) getCmCommandOutput(args ...string) string {
-	val, _ := p.env.runCommand("cm", args...)
+func (s *scm) getCmCommandOutput(args ...string) string {
+	val, _ := s.env.runCommand("cm", args...)
 	return val
 }
